# paypercrawl-backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: PayPerCrawl API
  description: Enterprise AI Bot Detection & Monetization Platform API
  version: 1.0.0
  contact:
    name: PayPerCrawl Support
    email: support@paypercrawl.tech
    url: https://paypercrawl.tech
  license:
    name: Proprietary
    url: https://paypercrawl.tech/terms
servers:
  - url: https://api.paypercrawl.tech/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /licenses/generate:
    post:
      tags: [Licenses]
      summary: Generate new license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [TRIAL, BASIC, PRO, ENTERPRISE]
                siteUrl:
                  type: string
                  format: uri
                maxSites:
                  type: integer
                  minimum: 1
                validDays:
                  type: integer
                  minimum: 1
      responses:
        201:
          description: License created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/License'
        400:
          $ref: '#/components/responses/BadRequest'
  /licenses/validate:
    post:
      tags: [Licenses]
      summary: Validate license key
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [licenseKey, siteUrl]
              properties:
                licenseKey:
                  type: string
                siteUrl:
                  type: string
                  format: uri
      responses:
        200:
          description: License validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  license:
                    $ref: '#/components/schemas/License'
        400:
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    License:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        type:
          type: string
        status:
          type: string
        maxSites:
          type: integer
        activeSites:
          type: integer
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
...